Implement memo create screen

+++ app/src/main/java/jp/co/cyberagent/dojo2020/data/model/Memo.kt
@@ -6,11 +6,11 @@ data class Memo(
     val id: String,
     val title: String,
     val contents: String,
-    val time: Double,
+    val time: Long,
     val category: String
 ) {
     companion object {
-        fun createWithOutId(title: String, contents: String, time: Double, category: String): Memo {
+        fun createWithOutId(title: String, contents: String, time: Long, category: String): Memo {
             val id = UUID.randomUUID().toString()
 
             return Memo(id, title, contents, time, category)
